feat(engine)!: separate into structure & style configs.
